feat: done all migrations

+++ src/main/resources/db/migration/V0002__create-table-user.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS tbl_user(
     user_id BIGSERIAL primary key,
     username varchar(255) NOT NULL UNIQUE,
     "password" varchar(255) NOT NULL,
-    role_id  bigint,
+    role_id  integer,
     CONSTRAINT fk_userid_roleid
         FOREIGN KEY (role_id)
             REFERENCES tbl_role(role_id)
